requirements.txt
streamlit
moviepy
requests
speechbrain
torchaudio
soundfile
pydub
gtts
numpy
Pillow
tensorflow
librosa
joblib